cleaned dwarf variable test cases, fixed problems
